server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.136.0",
    "fastapi-mail>=1.6.2",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "pydantic-settings>=2.14.0",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.26",
    "pytz>=2026.1.post1",
    "qrcode[pil]>=8.2",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.45.0",
    "groq>=1.2.0",
]
